requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
cryptography>=41.0.0
tiktoken>=0.5.0
pyotp>=2.9.0
qrcode>=7.4.0
pillow>=10.0.0
websockets>=12.0

# 性能优化
redis>=5.0.0
psutil>=5.9.0
aiofiles>=23.0.0
aiosqlite>=0.19.0

# RAG 向量检索
numpy>=1.24.0

# 企业功能（可选）
# ldap3>=2.9.0          # LDAP/AD 集成
# pinecone-client>=2.0.0  # Pinecone 向量数据库
# pymilvus>=2.3.0       # Milvus 向量数据库

# 支付集成（可选）
# stripe>=7.0.0         # Stripe 支付
# alipay-sdk-python>=3.0.0  # 支付宝

# 数据库（生产环境）
# asyncpg>=0.29.0       # PostgreSQL 异步驱动
# sqlalchemy>=2.0.0     # ORM

# 消息队列（可选）
# celery>=5.3.0         # 任务队列
# pika>=1.3.0           # RabbitMQ
# aiokafka>=0.9.0       # Kafka

# 监控与追踪
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0

# 安全增强
python-jose>=3.3.0    # JWT
passlib>=1.7.4        # 密码哈希
bcrypt>=4.1.0         # bcrypt 支持

# 时区支持
pytz>=2023.3

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
